requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
cffi==1.6.0
cryptography==1.3.2
dj-database-url==0.4.1
Django==1.9.6
django-extensions==1.6.6
djangorestframework==3.3.3
enum34==1.1.4
gunicorn==19.5.0
httplib2==0.9.2
idna==2.1
ipaddress==1.0.16
kombu==3.0.35
Markdown==2.6.6
ndg-httpsclient==0.4.0
oauth2==1.9.0.post1
oauthlib==1.1.1
psycopg2==2.6.1
pusher==1.2.3
pyasn1==0.1.9
pycparser==2.14
PyJWT==1.4.0
pyOpenSSL==16.0.0
python-openid==2.2.5
python-social-auth==0.2.19
pytz==2016.4
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
urllib3==1.15.1
Werkzeug==0.11.9
whitenoise==3.1
yelp==1.0.2
